Ported xf86-input-wacom. (Maybe)
xf86-input-wacom now uses XLibre as upstream and compiles.

That's half of the story, this codebase changed a lot form when the
original port was made.
The deleted patches are for files that are not even present in the
upstream except patch-test-wacom-tests.c this file was changed so much
that the patch didn't make sense.

The wcmValidateDevice.c file also has changed a lot and the old patch
completely fails, a new one should probably be written or a merge
request opened at the upstream with the required changes.
Hot plug may also not work.